(one) problem found, not fixed yet
We enable the tx interrupt. since there's nothing in the buffer, this
immediately triggers the interrupt handler, which gets up to
Msp430UsciUartP correctly. However, Msp430UsciUartP checks with the
arbiter to find the current client, and gets back 0xff rather than a
valid client. It then just returns, and as soon as the interrupt
routine exits, it gets re-entered (since the flag's not been cleared
by writing to the tx buffer).