flush RXBUF before starting a new read
this shouldn't really be necessary: if things work correctly, a read
operation ends by sending the NACK+STP in response to the last
expected byte.
However, we're starting the process of sending NACK + STP after the
last expected byte, which lets the slave send one more byte than
expected, which then ends up sitting in the RXBUF and making a mess of
things.