apparently safe implementation of slave/discovery
there appeared to be some race condition in the low-level I2C code,
though it may have just been an issue with the blocking printf's
interfering. The final byte was not being sent when the master read
from the slave, and the slave didn't get the STOP condition or the
final xmit request sometimes.

I wasn't able to pin down exactly where/why it was happening, but this
commit seems stable.