failed attempt to use BSLSKEY
incorrect password appears to erase all, still.

Not really sure what's not working here.

1.  Ideally, the binaries we generate will initialize BSLSKEY to
0x0000 automatically.  The sfrw macro usage in platform.h didn't work
(throws a syntax error when expanded). This should look something like
how the .vectors section is initialized.

2. Failing that, I tried to do this ugly maneuver in PlatformP that
clears it on boot. That didn't appear to work: I installed as usual,
then used install.sh with the wrong password, then plugged the board
back in. The previously-installed test app didn't run. argh. I think
that this is because I'm trying to write to main flash while the code
is executing in main flash, which is a no-no (has to execute from RAM
to modify main flash, i think).

A hacky workaround would be to modify the .ihex file: add a line
before the interrupt vector definition that explicitly initializes
this to 0.