only check frameType during PFS.
This should simplify the logic quite a bit. We end up needing to
switch from IDLE to FSTXON or RX during PFS, but We basically have to
do this anyway.  If we were able to always predict the next-frame
setup, then we could schedule PFS for later and only ask for frame
type if we are in IDLE during PFS. Not sure that this is a huge gain.