increased PFS_SLACK to resolve out-of-order exec
FSA was firing before the radio could be configured pretty frequently.
Really, we should be OK to do the radio config inline with PFS.f:
the tighter that we can make the timing between the radio config and
the start of the frame, the more energy-efficient things will be. As
long as it's being configured in the task context, it's going to be
hard to guarantee that this gets done in time, and so we're going to
need to add more slack (wasting energy).