adjust timings, use periodic alarm
When I used each tx sfd to resynchronize, nodes drifted apart over
several frames. When I use the first SFD only (and use a periodic
alarm over top of that), the drift is much less.

- original approach: first TX is in synch, each subsequent one the
  forwarder drifts ahead by ~2 uS (confirmed with multiple nodes)
- this approach: first TX is in synch, origin drifts ahead by ~1.5 uS
  over 5 total transmissions.

This is probably good enough. The next step would be to let the
original alarm fire, and then set a new very small alarm and poll the
IFG. This would ameliorate most of the interrupt handling jitter (if
there is any).