revert invFreq for router-subnetwork
the meaning of this field was being overloaded in some dumb ways. It
is used to set the wakeup period length, so if "invFreq=0 -> don't
probe", then routers will not probe to their subnetwork (prevents
multiple masters in same network) but routers will wait for 0 ticks
for their subnetwork to wake up.