increase TX_SLACK, fix frame timer usage
Left behind a TODO: we set the frame timer 5 frame-lengths after the
initial CTS. This accounts for the fact that the CTS message may take
multiple frames to finish.  It would be nicer to compute this, either
based on the time that the rx finished and the message timestamp or
based on the current time during the receive/handleCTS step. Either
way, perhaps just being conservative is the way to go.